requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
Django==5.1.2
django-cors-headers==4.5.0
django-geoposition==0.3.0
django-leaflet==0.31.0
django-parler==2.3
djangorestframework==3.15.2
fiona==1.10.1
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
networkx==3.4.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pip-review==1.3.0
postgis==1.0.4
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pygrib==2.1.6
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
shapely==2.0.6
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
